+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
   where `NotReady,SchedulingDisabled` would have red `NotReady`
   and orange `SchedulingDisabled`. (#69)
 
+- Added coloring on PV/StorageClass reclaim policy and PV/PVC status. (#73)
+
 - Changed to Go 1.21.5 to resolve vulnerability [GO-2023-2182](https://pkg.go.dev/vuln/GO-2023-2182),
   for denial of service in net/http, and [GO-2023-2185](https://pkg.go.dev/vuln/GO-2023-2185)
   for insecure parsing of Windows paths with a `\??\` prefix. (#72)